+++ packages/sampler/.storybook/main.js
@@ -4,8 +4,9 @@ module.exports = {
 	stories: ['./../stories/default/*.js'],
 	addons: [
 		'@enact/storybook-utils/addons/actions/register',
-		'@enact/storybook-utils/addons/knobs/register',
-		'@enact/storybook-utils/addons/docs/register'
+		'@enact/storybook-utils/addons/docs/register',
+		'@storybook/addon-controls',
+		'@storybook/addon-toolbars'
 	],
 	webpackFinal: async (config, {configType}) => {
 		return webpack(config, configType, __dirname);