+++ packages/sampler/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The following is a curated list of changes in the Enact Sampler, newest changes on the top.
 
+### Changed
+
+- Deprecated plugin `addon-knobs` replaced with `addon-controls`
+
 ## [4.0.3] - 2021-07-02
 
 No significant changes.